Subsequence Again (Hacker Rank)
You are provided with a string s and an integer k. You have to find another string t which satisfies the following conditions:

1: t must be a subsequence of s.

2: Every character in t must occur at least k times. 

3: The length of t must be as large as possible.

4: If there are multiple strings for t with largest possible length, pick the lexicographically smallest one.

For example, let's say the string is s = hackerrank and k = 2. 
The solution for this is akrrak. Here t is a subsequence of s, it contains the characters a,k and r repeated at least k times. 
And, it is the only longest possible subsequence that satisfies the conditions.