Mathison and pangrams (Code Chef)
Mathison recently inherited an ancient papyrus that contained some text. Unfortunately, the text was not a pangram. Now, Mathison has a particular liking for holoalphabetic strings and the text bothers him. The good news is that Mathison can buy letters from the local store in order to turn his text into a pangram. 
However, each letter has a price and Mathison is not very rich. Can you help Mathison find the cheapest way to obtain a pangram?