Sum of two numbers represented as arrays.py
Given two numbers represented by two arrays, write a function that returns sum array. The sum array is an array representation of addition of two input arrays. It is not allowed to modify the arrays.

Input:

The first line of input contains an integer T denoting the number of test cases.
The first line of each test case contains two integers M and N separated by a space. M is the size of arr1 and N is the size of arr2.
The second line of each test case contains M integers which are the input for arr1.
The third line of each test case contains N integers which are the input for arr2.

Output:

Print the sum list.